// src/modules/merchandising/pages/purchaseOrderEntry/purchaseOrderEntry.interface.ts
import { PaginationObject } from "@/types/global";

export interface IPurchaseOrder {
    id: number;
    factoryId: number;
    pono: string;
    buyerId: number;
    merchandiserId?: number;
    dealMerchandiserId?: number;
    followMerchandiserId?: number;
    agentId?: number;
    orderType?: string;
    extraCutPercent?: number;
    cuttingPaymentPercent?: number;
    sewingPaymentPercent?: number;
    finishingPaymentPercent?: number;
    packPaymentPercent?: number;
    totalPoQty?: number;
    totalProductionQty?: number;
    totalShipQty?: number;
    poReceiveDate?: Date;
    poConfirmDate?: Date;
    isActive: string;
    isSubContact?: string;
    seasonId?: number;
    brandId?: number;
    sectionNo?: string;
    kimballNo?: string;
    sapPoNo?: string;
    advanceReceive?: number;
    wagesAmount?: number;
    wagesAmount2?: number;
    wagesAmount3?: number;
    wagesDescription?: string;
    wagesDescription2?: string;
    wagesDescription3?: string;
    piNo?: string;
    piDate?: Date;
    piExpiryDate?: Date;
    lcBankId?: number;
    beniBankId?: number;
    shipTermId?: number;
    applicantId?: number;
    paymentTermId?: number;
    shipMode?: string;
    lcNo?: string;
    etdDate?: Date;
    poLId?: number;
    poDId?: number;
    consigneeId?: number;
    notifyId?: number;
    currencyId?: number;
    currencyRate?: number;
    exportLicence?: string;
    gsp?: string;
    remarks?: string;
    revisedNo?: string;
    isPlan?: number;
    additionalBookingId?: number;
    sampleProgramId?: number;
    createBy?: string;
    createDate?: Date;
    updateBy?: string;
    updateDate?: Date;
    isMultiPoCombine?: string;
    combinedJobNoId?: number;
    actionGuidId?: string;
}

export interface IPurchaseOrderState {
    purchaseOrder: IPurchaseOrder;
    purchaseOrders: IPurchaseOrder[];
    filteredBuyers: IPurchaseOrder[];
    paginationObject: PaginationObject<IPurchaseOrder>;
    loading: boolean;
    // validationErrors: BuyerValidationErrors | null;
    error: string | null;
    status: 'idle' | 'loading' | 'succeeded' | 'failed';
    message: string | null;
}